bench: isolate AddOverflow dispatch benchmark
